Get rid of some compilation warnings on Windows
Also remove DllMain() hack to set library search path during runtime. It
just does not work. Carla seems to be setting this path before loading
plugins.

+++ src/log.h
@@ -33,7 +33,7 @@
 #include <stdio.h>
 
 #define LOG_STDERR(msg)        fprintf(stderr, "%s : %s\n", __PRETTY_FUNCTION__, msg);
-#define LOG_STDERR_INT(msg,n)  fprintf(stderr, "%s : %s - %x\n", __PRETTY_FUNCTION__, msg, n);
+#define LOG_STDERR_INT(msg,n)  fprintf(stderr, "%s : %s - %x\n", __PRETTY_FUNCTION__, msg, (unsigned int)n);
 #define LOG_STDERR_ERRNO(msg)  fprintf(stderr, "%s : %s - %s\n", __PRETTY_FUNCTION__, msg, strerror(errno));
 #define LOG_STDERR_COLOR(msg)  fprintf(stderr, "\x1b[31m%s : %s\x1b[0m\n", __PRETTY_FUNCTION__, msg);
